feat: clean up CLI
