+++ src/spectre_cli/commands/delete.py
@@ -4,8 +4,8 @@
 
 import typer
 
-from spectre_cli.commands import safe_request
-from spectre_cli.commands import CliHelp
+from ._cli_help import CliHelp
+from ._safe_request import safe_request
 
 delete_app = typer.Typer(
     help = "Delete resources."
