+++ src/spectre_cli/commands/start.py
@@ -4,8 +4,8 @@
 
 import typer
 
-from spectre_cli.commands import safe_request
-from spectre_cli.commands import CliHelp
+from ._cli_help import CliHelp
+from ._safe_request import safe_request
 
 start_app = typer.Typer(
     help = "Start a job."
