+++ src/spectre_cli/commands/update.py
@@ -5,8 +5,8 @@
 import typer
 from typing import List
 
-from spectre_cli.commands import safe_request
-from spectre_cli.commands import CliHelp
+from ._cli_help import CliHelp
+from ._safe_request import safe_request
 
 update_app = typer.Typer(
     help = "Update resources."
